+++ frappe/integrations/doctype/nextcloud_settings/nextcloud_settings.js
@@ -2,9 +2,14 @@
 // For license information, please see license.txt
 
 frappe.ui.form.on('NextCloud Settings', {
+
 	refresh: function(frm) {
 		frm.clear_custom_buttons();
 		frm.events.take_backup(frm);
+	},
+
+	enabled: function(frm) {
+		frm.refresh();
 	},
 
 	take_backup: function(frm) {
